% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_Checker.R
\name{BoolBioNet_Checker}
\alias{BoolBioNet_Checker}
\title{Check the Boolean network is appropriate}
\usage{
BoolBioNet_Checker(OneBioNet)
}
\arguments{
\item{OneBioNet}{a 4-element formatted list, record Boolean network information}
}
\value{
a logical value, \code{TRUE}, for other downstream analysis. 
Any abnormal results and invalid inputs return error messages.
}
\description{
It compares and analyzes the lengths of node's names, input/output links, and Boolean function configuration. 
Check the Boolan network is structurally appropriate. For example, one node has three predecessors with 4-bit truth table is invalid.
}
\examples{
# Analysis of one real Boolean genetic network.
# BoolBioNet_Checker(BoolGRN_CellCollective$c_1557);
# Return TRUE (Denote all lists are appropriate)
BoolBioNet_Checker(BoolGRN_CellCollective$c_1557)
}
