% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_NodeAttr.R
\name{BoolBioNet_NodeAttr}
\alias{BoolBioNet_NodeAttr}
\title{Display a node's attribute}
\usage{
BoolBioNet_NodeAttr(RealBioNet, NodeName)
}
\arguments{
\item{RealBioNet}{A four-element formatted list that records information of 
a Boolean network.}

\item{NodeName}{Name of the node (gene).}
}
\value{
The information of the specified node.
}
\description{
This function returns the predecessors (inputs), successors (outputs), 
and regulatory patterns (Boolean function) of the specified nodes.
}
\examples{
# Print the information of the node "Akt" in the gene network [c_11863] from
# the CellCollective Set (BoolGRN_CellCollective).
# BoolBioNet_NodeAttr(BoolGRN_CellCollective$c_11863,"Akt");
# Return information:
# $TruthTable
#      PI3K [Akt]
# [1,]    0     0
# [2,]    1     1
# $Successor
# [1] "IKK"  "mTOR"
BoolBioNet_NodeAttr(BoolGRN_CellCollective$c_11863,"Akt")
}
