% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_NodeAttribute.R
\name{BoolBioNet_NodeAttribute}
\alias{BoolBioNet_NodeAttribute}
\title{Show a node's attribute}
\usage{
BoolBioNet_NodeAttribute(RealBioNet, NodeName)
}
\arguments{
\item{RealBioNet}{a 4-element formatted list, record biological network information.}

\item{NodeName}{Node/Gene's name.}
}
\value{
Attribute/information/feature of the node.
}
\description{
This function returns the predecessors, successors, and 
regulatory patterns of given nodes.
}
\examples{
# Print the information of Node "Akt" in gene network [c_11863] of 
# CellCollective Set.
# BoolBioNet_NodeAttribute(BoolGRN_CellCollective$c_11863,"Akt")
# Return information:
# $TruthTable
#      PI3K [Akt]
# [1,]    0     0
# [2,]    1     1
# $Successor
# [1] "IKK"  "mTOR"
BoolBioNet_NodeAttribute(BoolGRN_CellCollective$c_11863,"Akt")
}
