% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolBioNet_StrConComp.R
\name{BoolBioNet_StrConComp}
\alias{BoolBioNet_StrConComp}
\title{Caluclate the network's strongly connected components}
\usage{
BoolBioNet_StrConComp(OneNet)
}
\arguments{
\item{OneNet}{A four-element formatted list that records information of a Boolean network.}
}
\value{
A list in which each element represents a strongly connected component.
Where those labeled as \code{_TRUE} in the name are non-trivial SCCs, while 
\code{_FALSE} indicates terminal or relay nodes.
}
\description{
This function provides a brief analysis of the strongly connected 
components to facilitate comparison in core dynamic analysis.
}
\examples{
# Analyze the strongly connected components of a network.
# Here is an example of the network [c_2202] from the CellCollective Set
# BoolBioNet_StrConComp(BoolGRN_CellCollective$c_2202)
# Return a List
# [[1]]scc_1_FALSE   <-- a terminal node
# [1] "Exocytosis"
# [[2]]scc_2_FALSE   <-- a relay node 
# [1] "Packaging_Proteins"
# [[3]]scc_3_TRUE <-- a "real" SCC that meets the requirement of signal feedback.
# [1] "Protein_Phosphatase_1" "DARPP32"               "Calcineurin"          
# [4] "Calcium"               "Glutamate_Receptor"
# ... ... 

BoolBioNet_StrConComp(BoolGRN_CellCollective$c_2202)

}
