% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolFun_NestedCanalized.R
\name{BoolFun_NestedCanalized}
\alias{BoolFun_NestedCanalized}
\title{Decode the structure of a nested canalized Boolean function}
\usage{
BoolFun_NestedCanalized(Bit_Vec, PrintOut = FALSE)
}
\arguments{
\item{Bit_Vec}{a bool (or 0/1-integral) vector, represents a truth table of Boolean function.}

\item{PrintOut}{a bool value, show output to the terminal (Default: \code{FALSE}).}
}
\value{
a matrix that descirbes the nested structure of a canalized function.
}
\description{
A nested canalized Boolean function usually has a "if else-if ..."
regulating patterns. This function decodes the function to show the pattern 
intuitively.
}
\examples{
# Show the canalized structure of a nested canalized function.
# set.seed(202501)
# aboolfun<-BoolFun_Generator("C", VarNum=7L, vars=c(4L,9L))
# df_boolfun<-BoolFun_NestedCanalized(aboolfun, PrintOut=TRUE)
# Return a Matrix, and show nested "if-else-if" hierarchical relations.
set.seed(202501)
aboolfun<-BoolFun_Generator("C", VarNum=7L, vars=c(4L,9L))
df_boolfun<-BoolFun_NestedCanalized(aboolfun, PrintOut=TRUE)

}
