% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_B_NestedCanalized}
\alias{c_B_NestedCanalized}
\title{Decode the structure of a nested canalized Boolean function}
\usage{
c_B_NestedCanalized(boolfuns, k)
}
\arguments{
\item{boolfuns}{[\code{IntegerVector &}] the Boolean function logical vector.}

\item{k}{[\code{int}] the number of input variables of the function.}
}
\description{
C++ prototype function of \link{BoolFun_NestedCana}.
}
