% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_Generator}
\alias{c_BF_Generator}
\title{Generate a specific type of Boolean function}
\usage{
c_BF_Generator(Type, k, bias, Vars)
}
\arguments{
\item{Type}{[\code{char}] function type.}

\item{k}{[\code{k}] the number of input variables of the function.}

\item{bias}{[\code{double}] function bias.}

\item{Vars}{[\code{IntegerVector}] some configuration parameters.}
}
\description{
C++ prototype function of \link{BoolFun_Generator}.
}
