% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_isPointed}
\alias{c_BF_isPointed}
\title{C++ prototype function for BoolFun_Type()}
\usage{
c_BF_isPointed(boolfunss, k, Leixing, Showit)
}
\arguments{
\item{boolfunss, }{bool function logical vector.}

\item{k, }{input variable size.}

\item{Leixing, }{function's type.}

\item{Showit, }{whether show the function?}
}
\description{
C++ prototype function for BoolFun_Type()
}
\details{
Please see document of \link{BoolFun_Type}
}
