% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_BF_QuineMcCluskey}
\alias{c_BF_QuineMcCluskey}
\title{Display the Quine-McCluskey form of a Boolean function}
\usage{
c_BF_QuineMcCluskey(boolfuns, k, VarsName)
}
\arguments{
\item{boolfuns}{[\code{IntegerVector &}] the Boolean function.}

\item{k}{[\code{int}] the number of input variables of the function.}

\item{VarsName}{[\code{CharacterVector}] variable names.}
}
\description{
C++ prototype function of \link{BoolFun_QMForm}.
}
