% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_CoreDynamicNode}
\alias{c_CoreDynamicNode}
\title{Analyze core dynamic components of Boolean networks (1)}
\usage{
c_CoreDynamicNode(
  aRealNet,
  PointedGene,
  PointValues,
  InD,
  OtD,
  NodeDetailInfor,
  ReturnResidualNetwork,
  Times
)
}
\arguments{
\item{aRealNet}{[\code{List}] a real/silico genetic network (See \link{BoolGRN_CellCollective}).}

\item{PointedGene}{[\code{IntegerVector}] which nodes should be fixed.}

\item{PointValues}{[\code{IntegerVector}] pointed values of free/controlled nodes.}

\item{InD}{[\code{IntegerVector}] the vector of in-degrees.}

\item{OtD}{[\code{IntegerVector}] the vector of out-degrees.}

\item{NodeDetailInfor}{[\code{int}] should return detail information of node (>0, yes).}

\item{ReturnResidualNetwork}{[\code{int}] should return Residual Network (>0, yes).}

\item{Times}{[\code{int}] recursive number of analysis (>0).}
}
\description{
C++ prototype function of \link{BoolBioNet_CoreDyn} (\code{coredyn} part).
}
