% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_Derrida_Simualtion}
\alias{c_Derrida_Simualtion}
\title{Analyze damage spread within system}
\usage{
c_Derrida_Simualtion(
  sys_size,
  ll_system,
  sim_step,
  obf_type,
  bias_rf,
  obf_ratio,
  init_dis,
  init_1_ratio,
  net_type,
  net_f_para,
  obf_i_para1,
  obf_i_para2,
  RuleType
)
}
\arguments{
\item{sys_size}{[\code{int}] system size}

\item{ll_system}{[\code{int}] level number of discrete system}

\item{sim_step}{[\code{int}] simulating steps}

\item{obf_type}{[\code{char}] ordered Boolean function type}

\item{bias_rf}{[\code{NumericVector}] bias in random function}

\item{obf_ratio}{[\code{double}] proportion of ordered Boolean function}

\item{init_dis}{[\code{double}] initial normalized Hamming distance}

\item{init_1_ratio}{[\code{NumericVector}] "1" proportion of initial state}

\item{net_type}{[\code{char}] system topological type}

\item{net_f_para}{[\code{double}] topological configured parameters}

\item{obf_i_para1}{[\code{int}] configuration parameter for OBF (1)}

\item{obf_i_para2}{[\code{int}] configuration parameter for OBF (2)}

\item{RuleType}{[\code{int}] update strategy}
}
\description{
C++ prototype function of \link{DNS_DamageSpread}
}
