% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_M_Domainted}
\alias{c_M_Domainted}
\title{C++ prototype function for MulVFun_is_Domainted()}
\usage{
c_M_Domainted(amulfunss, k, L, Lens)
}
\arguments{
\item{amulfunss, }{a multi-valued function vector.}

\item{k, }{input variable size.}

\item{L, }{L-level discrete value.}

\item{Lens, }{length of mapping table: L^k.}
}
\description{
C++ prototype function for MulVFun_is_Domainted()
}
\details{
Please see document of \link{MulVFun_is_Domainted}
}
