% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulVF_Generator}
\alias{c_MulVF_Generator}
\title{Generate a specific type of multi-valued function}
\usage{
c_MulVF_Generator(
  FunType,
  k,
  L,
  CanaDeep,
  CanaVar,
  CanaVarNum,
  CanaInfo1,
  CanaInfo2,
  bias,
  Cana_Free
)
}
\arguments{
\item{FunType}{[\code{char}] function type.}

\item{k}{[\code{int}] the number of input variables.}

\item{L}{[\code{int}] the level of function.}

\item{CanaDeep}{[\code{int}] the layer of canalization.}

\item{CanaVar}{[\code{IntegerVector &}] canalized variable indexes.}

\item{CanaVarNum}{[\code{IntegerVector &}] the canalizing number of each variable.}

\item{CanaInfo1}{[\code{List &}] canalizing information.}

\item{CanaInfo2}{[\code{List &}] canalized information.}

\item{bias}{[\code{NumericVector &}] function bias.}

\item{Cana_Free}{[\code{bool}] is a parameter-free canalized function?}
}
\description{
C++ prototype function of \link{MulVFun_Generator}
}
