% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulVF_Sensitivity}
\alias{c_MulVF_Sensitivity}
\title{Calculate the sensitivity of a multi-valued function}
\usage{
c_MulVF_Sensitivity(amulfun, k, L, Lens)
}
\arguments{
\item{amulfun}{[\code{IntegerVector &}] the function.}

\item{k}{[\code{int}] the number of input variables.}

\item{L}{[\code{int}] the level of function.}

\item{Lens}{[\code{int}] the length of mapping table (\eqn{L^k}).}
}
\description{
C++ prototype function of \link{MulVFun_Sensitivity}
}
