% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulVFun2Polynomial}
\alias{c_MulVFun2Polynomial}
\title{C++ prototype function for MulVFun_Polynomial()}
\usage{
c_MulVFun2Polynomial(VariableMat, MapTab, k, L)
}
\arguments{
\item{VariableMat, }{a matrix that record input variables.}

\item{MapTab, }{the truth table of multi-valued function.}

\item{k, }{input variable size.}

\item{L, }{L-level discrete value.}
}
\value{
a List, [[1]] is.SAT ture or not? [[2]] Weights of coupled variable; [[3]] thresholds of L-levels.
}
\description{
C++ prototype function for MulVFun_Polynomial()
}
\details{
Please see document of \link{MulVFun_Polynomial}
}
