% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_Percolation_Simualtion}
\alias{c_Percolation_Simualtion}
\title{Analyze percolation within system}
\usage{
c_Percolation_Simualtion(
  sys_size,
  ll_system,
  sim_step,
  lat_type,
  obr_window,
  obf_type,
  bias_rf,
  obf_ratio,
  init_bias,
  net_type,
  net_f_para,
  obf_i_para1,
  obf_i_para2,
  OutPut,
  RuleType
)
}
\arguments{
\item{sys_size}{[\code{int}] system size}

\item{ll_system}{[\code{int}] level number of discrete system}

\item{sim_step}{[\code{int}] simulating steps}

\item{lat_type}{[\code{int}] lattice type}

\item{obr_window}{[\code{int}] observe window}

\item{obf_type}{[\code{char}] ordered Boolean function type}

\item{bias_rf}{[\code{NumericVector}] biases in random function}

\item{obf_ratio}{[\code{double}] proportion of ordered Boolean function}

\item{init_bias}{[\code{NumericVector}] "1" proportion of initial state}

\item{net_type}{[\code{char}] system topological type}

\item{net_f_para}{[\code{char}] topological configured parameters}

\item{obf_i_para1}{[\code{int}] configuration parameter for OBF (1)}

\item{obf_i_para2}{[\code{int}] configuration parameter for OBF (2)}

\item{OutPut}{[\code{bool}] whehther output all states?}

\item{RuleType}{[\code{int}] update strategy}
}
\description{
C++ prototype function of \link{DNS_Percolation}
}
