% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_ScalingLaw_RealNet}
\alias{c_ScalingLaw_RealNet}
\title{C++ prototype function for BoolBioNet_CoreDyn()}
\usage{
c_ScalingLaw_RealNet(
  aRealNet,
  PointedGene,
  PointValues,
  InD,
  OtD,
  NodeDetailInfor,
  ReturnResidualNetwork
)
}
\arguments{
\item{aRealNet, }{list-object of real/silico genetic network}

\item{PointedGene, }{integer or name vector, which nodes should be fixed.}

\item{PointValues, }{pointed values, for free or controlled nodes.}

\item{InD, }{in-degree vector.}

\item{OtD, }{out-degree vector.}

\item{NodeDetailInfor, }{should return detail informations of node.}

\item{ReturnResidualNetwork, }{should return Residual Network >0: yes.}
}
\description{
C++ prototype function for BoolBioNet_CoreDyn()
}
\details{
Please see document of \link{BoolBioNet_CoreDyn}
}
