% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_StrongConnectComponent}
\alias{c_StrongConnectComponent}
\title{C++ prototype function for BoolBioNet_StrConComp()}
\usage{
c_StrongConnectComponent(aRealNet, InD, OtD)
}
\arguments{
\item{aRealNet, }{list-object of real/silico genetic network}

\item{InD, }{IntegerVector of in-degree}

\item{OtD, }{IntegerVector of out-degree}
}
\value{
a List, each element is a strong connect component.
}
\description{
C++ prototype function for BoolBioNet_StrConComp()
}
\details{
Please see document of \link{BoolBioNet_StrConComp}
}
