% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBoolBioNetwork.R
\name{LoadBoolBioNetwork}
\alias{LoadBoolBioNetwork}
\title{Load a Boolean genetic network}
\usage{
LoadBoolBioNetwork(
  NetName,
  NetType = c("BoolExpression", "TruthTable", "Threshold")
)
}
\arguments{
\item{NetName}{a valid file/address of target GenNet.}

\item{NetType}{Which input format? Limited to \code{BoolExpression}, 
\code{TruthTable}, \code{Threshold}.
\itemize{
\item \code{BoolExpression}: all genes in a single file are shown as Boolean expression like,
\preformatted{
  X_1 = X_3 AND X_4
  X_2 = NOT (X_1 OR X_3)
  ... ...
}
\item \code{TruthTable}: all genes stored in a specified directory like (Based on organizational forms of Cell Collective Web),
\preformatted{
  in NetName/X_1.csv (for gene X_1)
     X_3, X_5, X_1
      0 ,  0 ,  0
      0 ,  1 ,  1
      1 ,  0 ,  1
      1 ,  1 ,  0
  in NetName/X_2.csv (for gene X_2)
     X_4, X_2
      0 ,  0
      1 ,  1
      ... ...
}
\item \code{Threshold}: represent all regulations 1 for activation, 2 for inhibition.
\preformatted{
     X_1  X_3  1
     X_4  X_2  2
     X_5  X_2  1
      ... ...
}
}
The detail files and organizational structures can see \href{https://cellcollective.org}{Cell Collective Website}}
}
\value{
a list contains member names of system, topological structure,
and Boolean function.
}
\description{
Load genetic network information from external files.
}
