% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadMulValBioNets.R
\name{LoadMulValBioNets}
\alias{LoadMulValBioNets}
\title{Load a multi-valued genetic network}
\usage{
LoadMulValBioNets(NetName)
}
\arguments{
\item{NetName}{A string that denotes a valid file path or name for the target 
network.}
}
\value{
A four-element formatted list that records information of a multi-valued
network. Note that the 4-th elements save weights rather than the Boolean funciton.
}
\description{
Load genetic network information from external files.
}
