% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulV2Bool_Bool2MulV.R
\name{MulV2Bool_Bool2MulV}
\alias{MulV2Bool_Bool2MulV}
\title{Boolean to multi-valued and multi-valued to Boolean transformation.}
\usage{
MulV2Bool_Bool2MulV(
  aVec,
  k = 0,
  L = 3,
  Thres = NA,
  Bool2MulV = TRUE,
  MappingTable = FALSE
)
}
\arguments{
\item{aVec}{IntegerVector, represents a mapping table of multi-valued function.}

\item{k}{an integer, k-variable of function.}

\item{L}{an integer, L-level multi-valued system (Default: 3).}

\item{Thres}{IntegerVector, Thresholds of multi-valued system. Its length is k+1;
each element is [1,L-1].}

\item{Bool2MulV}{a logical values, Original map is Boolean (Default: TRUE)}

\item{MappingTable}{logical value. Is the mapping table also returned? (Default: \code{FALSE})}
}
\value{
an IntegerVector or IntegerMatrix, mapping table of target system.
}
\description{
Boolean to multi-valued and multi-valued to Boolean transformation.
}
\examples{
# Transform a Boolean to a multi-valued one.
# the Boolean is f(x,y)=x & y, the binary threshold is [0,1), [1,3), see results
# set.seed(2025)
# MulV2Bool_Bool2MulV(c(0,0,0,1), 2, 3, c(1,1,1), MappingTable = TRUE)
# Return
#       v1 v0 xx
#  [1,]  0  0  0
#  [2,]  0  1  0
#  [3,]  0  2  0
#  [4,]  1  0  0
#  [5,]  1  1  2
#  [6,]  1  2  2
#  [7,]  2  0  0
#  [8,]  2  1  1
#  [9,]  2  2  2
set.seed(2025)
MulV2Bool_Bool2MulV(c(0,0,0,1), 2, 3, c(1,1,1), MappingTable = TRUE)

}
