% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_Polynomial.R
\name{MulVFun_Polynomial}
\alias{MulVFun_Polynomial}
\title{Convert a multi-valued function as polynomial like forms}
\usage{
MulVFun_Polynomial(aVec, k = 2, L = 3)
}
\arguments{
\item{aVec}{An IntegerVector that represents a mapping table of multi-valued 
function.}

\item{k}{An integer that denotes the number of input variables (Default: 2).}

\item{L}{An integer that represents the level of multi-valued system (Default: 3).}
}
\value{
A list containing three elements:
\itemize{
\item \code{[[1]] sat}: Is satisfiability? \code{TRUE} or \code{FALSE}
\item \code{[[2]] Wights}: The weights and thresholds. 
\item \code{[[3]] HighOrder}: The highest order of terms contained in the 
function (See examples).}
}
\description{
Unlike the Boolean system, we adopt a definition similar to 
that of domaineted function; please refer to \link{MulVFun_is_Domainted}.
}
\examples{
# Convert the a multi-valued function into a polynomial like expression.
# In fact, the result is same as MulVFun_is_Domainted().
# MulVFun_Polynomial(c(0,1,2,1,1,2,2,1,2), k=2L, L=3L);
# Return 
# [[1]] sat:TRUE, 
# [[2]] Weight:
#    x_1         x_2 threshold_1 threshold_2 threshold_3 
#      2           1          -2           0           0
# [[3]] HighOrder: 1

MulVFun_Polynomial(c(0,1,2,1,1,2,2,1,2), k=2L, L=3L)

}
