% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulVFun_QMForm.R
\name{MulVFun_QMForm}
\alias{MulVFun_QMForm}
\title{Show a multi-valued function's Quine-McCluskey form}
\usage{
MulVFun_QMForm(
  aVec,
  k,
  L,
  VarsName = NULL,
  ShowType = c("print", "data.frame"),
  SourceTable = FALSE
)
}
\arguments{
\item{aVec}{a vector, represents a truth table of valued function.}

\item{k}{an integer, k-variable of function.}

\item{L}{an integer, L-level multi-valued system.}

\item{VarsName}{a character vector, represents each variable's name}

\item{ShowType}{\code{c("print","data.frame")}, limited to \code{print} 
(Print in terminal) or \code{data.frame} (as a data.frame form).}

\item{SourceTable}{logical, should return the source mapping table (Default: 
\code{FALSE})? If \code{TRUE}, return a list: [[1]] QMC form, [[2]] source 
table. It is invalid when \code{ShowType} is \code{"print"}.}
}
\value{
dataframe/list/NULL, show the multi-valued QMC form.
}
\description{
The function employs our native C++ standard multi-valued 
Quine-McCluskey algorithm to obtain prime implicants. Because of algorithms 
and system properties, the result is not unique. See reference 
[\href{https://doi.org/10.1007/s00500-007-0175-x}{Petrik,2008}].
}
\examples{
# Calculate the Quine-McCluskey form of multi-valued function.
# Analyze a three-valued multiplexer: 
#             | a, iff x=0
# f(x,c,b,a)= | b, iff x=1
#             | c, iff x=2
# three_valued_Multiplexer; # Show the multiplexer.
# MulVFun_QMForm(three_valued_Multiplexer[,5], k=4,L=3,
#   VarsName=colnames(three_valued_Multiplexer)[4:1]);# Return the QMC form.
# Return: 
# x  c  b  a f(*)
# 2  1 NA NA    1
# 2  2 NA NA    2
# 1 NA  1 NA    1
# 1 NA  2 NA    2
# 0 NA NA  1    1
# 0 NA NA  2    2
MulVFun_QMForm(three_valued_Multiplexer[,5], k=4,L=3, 
   VarsName=colnames(three_valued_Multiplexer)[4:1]);
}
