% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_CheckValidMulVFun.R
\name{r_CheckValidMulVFun}
\alias{r_CheckValidMulVFun}
\title{Determines whether the input vector is a valid mapping result.}
\usage{
r_CheckValidMulVFun(a_Vec, k, L)
}
\arguments{
\item{a_Vec}{a integral/logical vector, represents a mapping table of multi-valued function.}

\item{k}{an integer, k-variable of function.}

\item{L}{an integer, L-level multi-valued system.}
}
\value{
IntegerVector
}
\description{
Determines whether the input vector is a valid mapping result.
}
