% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_GenerateTruthTable_M.R
\name{r_GenerateTruthTable_M}
\alias{r_GenerateTruthTable_M}
\title{Generate a truth table with pointed in-degree function.}
\usage{
r_GenerateTruthTable_M(inDegree = 2L, Lsystem = 2L)
}
\arguments{
\item{inDegree}{an integrate value, represents a truth table of Boolean function.}

\item{Lsystem}{an integer, L-level multi-valued system.}
}
\value{
a data.frame of input case enumeration
}
\description{
Generate a truth table with pointed in-degree function.
}
