% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_UnfreezeInputNode.R
\name{r_UnfreezeInputNode}
\alias{r_UnfreezeInputNode}
\title{Add self-loop circuits to all external nodes}
\usage{
r_UnfreezeInputNode(RealBioNet)
}
\arguments{
\item{RealBioNet}{A four-element formatted list that records information of a
Boolean network (See \link{BoolGRN_CellCollective})}
}
\value{
A four-element formatted list that records information of the modified
Boolean network (See \link{BoolGRN_CellCollective})
}
\description{
Convert all input and external nodes into self-loop nodes
}
