% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transient_Process_G.R
\name{Transient_Process_G}
\alias{Transient_Process_G}
\title{Simulating a transient process (Automatically generate network)}
\usage{
Transient_Process_G(
  Size = 1000L,
  SimStep = 1000L,
  Init_Dist = 0.1,
  Init_1_Ratio = c(0.5, 0.5),
  OBF_Type = "R",
  OBF_iPara1 = 1L,
  OBF_iPara2 = 1L,
  OBF_Ratio = 0.1,
  RBF_Bias = c(0.5, 0.5),
  Net_Type = "K",
  Net_fPara = 4,
  NumSys = 2L,
  UpdateRule = 1L
)
}
\arguments{
\item{Size}{an integer, size of system.}

\item{SimStep}{an integer, steps of simulating system.}

\item{Init_Dist}{a numeric value, initial normalized Hamming distance of two vector.}

\item{Init_1_Ratio}{a NumvericVector, proportion of each value (its length is \code{NumSys}).}

\item{OBF_Type}{a character, ordered Boolean function type (OBF).}

\item{OBF_iPara1}{an integer, configuration parameter for OBF.}

\item{OBF_iPara2}{an integer, configuration parameter for OBF (Not necessary for some types of OBF).}

\item{OBF_Ratio}{a numeric value, proportion of ordered Boolean function within system.}

\item{RBF_Bias}{a NumvericVector, biases of random function (each is (0,1), sum=1).}

\item{Net_Type}{a character, system topological type.}

\item{Net_fPara}{a numeric value, topological configured parameters.}

\item{NumSys}{an integer, number of discrete value.}

\item{UpdateRule}{an integer, update rule: 1=synchronous, 2=asynchronous, 
3=quick-asynchronous.}
}
\value{
A list containing two elements: 
\itemize{
\item \code{[[1]] Transient states}: A matrix whose rows and columns represent 
 time points and nodes, respectively.
\item \code{[[2]] Network}: The structure is consistent with that in 
\link{BoolGRN_CellCollective}
}
}
\description{
Same as \link{DNS_DamageSpread}, but record the transient process.
}
\details{
See \link{DNS_DamageSpread}.
}
