% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transient_Process_M.R
\name{Transient_Process_M}
\alias{Transient_Process_M}
\title{Simulating roughly attractor analysis of multi-valued model}
\usage{
Transient_Process_M(RealBioNet_M, SampleNum = 100L, Times = 1000L, NumSys = 2L)
}
\arguments{
\item{RealBioNet_M}{A four-element formatted list that records information of a multi-valued threshold network.}

\item{SampleNum}{An integer pointing the number of randomly sampling initial states.}

\item{Times}{an integer, steps of simulating system.}

\item{NumSys}{an integer, number of discrete value.}
}
\value{
A matrix where each row represents the final state of a stochastic 
simulation, and each column represents a gene.
}
\description{
Like \link{Transient_Process_A}, but use multi-valued model.
}
\details{
See the reference [\href{https://doi.org/10.1186/1752-0509-1-4}{Schaub2007}]
}
