% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_B_NestedCanalized}
\alias{c_B_NestedCanalized}
\title{C++ prototype function for BoolFun_NestedCanalized()}
\usage{
c_B_NestedCanalized(boolfunss, k)
}
\arguments{
\item{boolfunss, }{truth table.}

\item{k, }{input variable size.}
}
\description{
C++ prototype function for BoolFun_NestedCanalized()
}
\details{
Please see document of \link{BoolFun_NestedCanalized}
}
