% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_Derrida_Simualtion}
\alias{c_Derrida_Simualtion}
\title{C++ prototype function for DNS_DamageSpread()}
\usage{
c_Derrida_Simualtion(
  sys_size,
  ll_system,
  sim_step,
  obf_type,
  bias_rf,
  obf_ratio,
  init_dis,
  init_1_ratio,
  net_type,
  net_f_para,
  obf_i_para1,
  obf_i_para2,
  RuleType
)
}
\arguments{
\item{sys_size, }{system size}

\item{ll_system, }{level number of discrete system}

\item{sim_step, }{simulating steps}

\item{obf_type, }{ordered Boolean function type}

\item{bias_rf, }{biases in random function}

\item{obf_ratio, }{proportion of ordered Boolean function}

\item{init_dis, }{initial normalized Hamming distance}

\item{init_1_ratio, }{"1" proportion of initial vector}

\item{net_type, }{system topological type}

\item{net_f_para, }{topological configured parameters}

\item{obf_i_para1, }{configuration parameter for OBF (1)}

\item{obf_i_para2, }{configuration parameter for OBF (2)}

\item{RuleType, }{update strategy}
}
\description{
C++ prototype function for DNS_DamageSpread()
}
\details{
Please see document of \link{DNS_DamageSpread}
}
