% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_MulVF_Generator}
\alias{c_MulVF_Generator}
\title{C++ prototype function for MulVFun_Generator()}
\usage{
c_MulVF_Generator(
  FunType,
  k,
  L,
  CanaDeep,
  CanaVar,
  CanaVarNum,
  CanaInfo1,
  CanaInfo2,
  bias,
  Cana_Free
)
}
\arguments{
\item{FunType, }{function's type.}

\item{k, }{input variable size.}

\item{L, }{level of discrete system.}

\item{CanaDeep, }{layer of canalization.}

\item{CanaVar, }{Canalized varibale's ID.}

\item{CanaVarNum, }{Each canalizing number.}

\item{CanaInfo1, }{Rcpp::List, canalizing information.}

\item{CanaInfo2, }{Rcpp::List, canalized information.}

\item{bias, }{function's bias.}

\item{Cana_Free, }{is a parameter-free canalized function?}
}
\description{
C++ prototype function for MulVFun_Generator()
}
\details{
Please see document of \link{MulVFun_Generator}
}
